Omit BEGIN/COMMIT statements for empty transactions
If a transaction is opened and closed without any queries being run, we
can safely omit the `BEGIN` and `COMMIT` statements, as they only exist
to modify the connection's behaviour inside the transaction. This
removes the overhead of those statements when saving a record with no
changes, which makes workarounds like `save if changed?` unnecessary.

This implementation wraps the raw database connection in a proxy which
buffers `begin_transaction` calls, and materializes them the next time
the connection is used, or discards them if `end_transaction` is called
first. This approach makes new connection usage inside the adapters
behave correctly by default. Third party adapters can use the
`begin_transaction`/`end_transaction` APIs to opt in to the new
behaviour, but will still work even if they don't.

If `raw_connection` is used to get a reference to the underlying
database connection, the behaviour is disabled and transactions are
opened eagerly, as we can't know how the connection will be used.
However when the connection is checked back into the pool, we can assume
that the application won't use the reference again and reenable lazy
transactions. This prevents a single `raw_connection` call from
disabling lazy transactions for the lifetime of the connection.