Only invalidate query caches for peer connections
Currently writes invalidate the query cache for all connections that are
owned by the current thread, but this causes unnecessary cache misses
when reading from one database and writing to another.

We only need to ensure that writes can be read from their corresponding
reading connections, so only clearing the cache for connections with the
same `connection_specification_name` is sufficient.